update dan fix error

+++ Zia.h
@@ -25,17 +25,17 @@ void pertambahan_kurang(){
 	}
 }
 
-int perjumlahan(int bil1, int bil2){
+double perjumlahan(double bil1, double bil2){
 	return  bil1 + bil2;
 }
-int pengurangan(int bil1, int bil2){
+double pengurangan(double bil1, double bil2){
 	return bil1 - bil2;
 }
-int perkalian(int bil1, int bil2){
+double perkalian(double bil1, double bil2){
 	return bil1 * bil2;
 }
 
-int pembagian(int bil1, int bil2){
+double pembagian(double bil1, double bil2){
 	return bil1 / bil2;
 }
